BugFix: Error in CandidateSet for m<n
In case that the number m of candidates per node is less than the total
number of nodes n, i.e., the common case, the CandidateSet will be
wrong. This is a very serious issue, as it would likely have very
negative influence on the algorithm's performance. The reason was a
very stupid one: I populate the list from the front, while my node
counter comes from the back. A simple change in iteration order fixed
it.